poc/requirements.txt
pyramid
WebTest
nose
deforms
dnspython
easyzone
